docs: finish main readme.md. cleanup examples
